Nextlist

+++ src/CurListsPage/CurListsPage.js
@@ -32,20 +32,14 @@ class CurListsPage extends Component {
   }
   deleteList = (listId) => {
     const { select ='' } = this.props;
-    if (select === "Now"){
-      BuyListApiService.deleteBuyList(listId)
+
+    
+    BuyListApiService.deleteBuyList(listId)
       .then(res => {
-        this.context.deleteBuyList(listId);
+        if (select === "Now")this.context.deleteBuyList(listId);
+        else this.context.deleteNextList(listId);
       })
       .catch(this.context.setError)
-    }
-    else {
-      BuyListApiService.deleteNextList(listId)
-      .then(res => {
-        this.context.deleteNextList(listId);
-      })
-      .catch(this.context.setError)
-    }
     
   }
   changeButtonClick = (listId) => {
@@ -64,28 +58,41 @@ class CurListsPage extends Component {
     ev.preventDefault();
     const {updateList} = ev.target;
     const { select ='' } = this.props;
-    if (select === 'Now') {
-      BuyListApiService.updateBuyList(this.state.selectedListId, updateList.value)
-      .then(res => {
-        this.context.updateBuyList(this.state.selectedListId, updateList.value)
-        this.setState({
-          textAreaActive: false,
-          selectedListId: null,
-        })
+    // this.context.updateNextList(this.state.selectedListId, updateList.value)
+    BuyListApiService.updateBuyList(this.state.selectedListId, updateList.value)
+    .then(res => {
+      console.log(select);
+      console.log(updateList.value);
+      if (select === 'Now') this.context.updateBuyList(this.state.selectedListId, updateList.value)
+      else this.context.updateNextList(this.state.selectedListId, updateList.value)
+      this.setState({
+        textAreaActive: false,
+        selectedListId: null,
       })
-      .catch(err => this.context.setError(err.error))
-    }
-    else{
-      BuyListApiService.updateNextList(this.state.selectedListId, updateList.value)
-      .then(res => {
-        this.context.updateNextList(this.state.selectedListId, updateList.value)
-        this.setState({
-          textAreaActive: false,
-          selectedListId: null,
-        })
-      })
-      .catch(err => this.context.setError(err.error))
-    }
+    })
+    .catch(err => this.context.setError(err.error))
+    // if (select === 'Now') {
+    //   BuyListApiService.updateBuyList(this.state.selectedListId, updateList.value)
+    //   .then(res => {
+    //     this.context.updateBuyList(this.state.selectedListId, updateList.value)
+    //     this.setState({
+    //       textAreaActive: false,
+    //       selectedListId: null,
+    //     })
+    //   })
+    //   .catch(err => this.context.setError(err.error))
+    // }
+    // else{
+    //   BuyListApiService.updateNextList(this.state.selectedListId, updateList.value)
+    //   .then(res => {
+    //     this.context.updateNextList(this.state.selectedListId, updateList.value)
+    //     this.setState({
+    //       textAreaActive: false,
+    //       selectedListId: null,
+    //     })
+    //   })
+    //   .catch(err => this.context.setError(err.error))
+    // }
   }
   renderLists(selectLists, select) {
     return selectLists.map(list =>
